requirements.txt
duolingo-api==0.5.3
requests~=2.25.0
python-dotenv~=0.15.0